requirements.txt
typing==3.5
fastapi==1.8.2
uvicorn==0.14.0
starlette==0.14.0
python-multipart
sqlalchemy==1.4.20
psycopg2==2.9.1
databases
python-dotenv==0.18.0
python-jose==3.3.0
passlib==1.7.4
pydantic[email]
bcrypt==3.2.0
alembic
pydantic==3.10.0.0
asyncpg==0.23.0
jinja2==3.0.1